requirements.txt
absl-py==2.0.0
aiohttp==3.9.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
arviz==0.16.1
astropy==5.3.4
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
cloudpickle==2.2.1
cmake==3.27.5
comm==0.1.4
contourpy==1.1.1
corner==2.2.2
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
einops==0.6.1
emcee==3.1.4
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.12.4
fonttools==4.42.1
frozenlist==1.4.1
fsspec==2024.2.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
grpcio==1.58.0
h5netcdf==1.2.0
h5py==3.9.0
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
lightning-utilities==0.10.1
lit==16.0.6
llvmlite==0.41.0
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.5.8
networkx==3.1
nflows==0.14
numba==0.58.0
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
pluggy==1.3.0
prompt-toolkit==3.0.39
protobuf==4.24.3
provabgs @ git+https://github.com/changhoonhahn/provabgs.git@fead507d3ef04d117b70b736208e5cbf9bae29ff
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyerfa==2.0.0.3
Pygments==2.16.1
pyknos==0.15.1
pyparsing==3.1.1
pyro-api==0.1.2
pyro-ppl==1.8.6
pytest==7.4.2
pytest-astropy-header==0.2.2
python-dateutil==2.8.2
pytorch-lightning==2.2.0.post0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sbi==0.21.0
SciencePlots==2.1.0
scikit-learn==1.3.1
scipy==1.11.2
seaborn==0.13.0
six==1.16.0
speclite==0.17
stack-data==0.6.2
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensordict==0.1.2
threadpoolctl==3.2.0
tomli==2.0.1
torch==2.0.1
torchmetrics==1.3.1
torchrl==0.1.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.7
xarray==2023.8.0
xarray-einstats==0.6.0
yarl==1.9.4
zeus-mcmc==2.5.4
